renamed definition to adapterdef
