+++ python/tests/test_integration_essential.py
@@ -8,7 +8,7 @@ import pytest
 from pathlib import Path
 from click.testing import CliRunner
 
-from treeviz.cli import cli
+from treeviz.__main__ import cli
 from treeviz.formats import parse_document
 from treeviz.adapters import adapt_node
 from treeviz.renderer import render
